requirements.txt
streamlit
requests
feedparser
newspaper3k
nltk
gtts
python-dotenv
dateutil
urllib3
